# docs/contracts/authoring-public-v1.openapi.yaml
openapi: 3.0.3
info:
  title: Engineering Authoring Coach API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:28083
security:
  - bearerAuth: []
paths:
...
  /api/v1/courses/{courseId}:
    get:
      summary: Read a published course without internal content identifiers
      parameters: [{$ref: '#/components/parameters/courseId'}]
      responses: {'200': {description: Student course details}}
...
  /api/v1/admin/courses:
    get:
      summary: List courses for administration
      responses: {'200': {description: Courses}, '403': {description: ADMIN role required}}
    post:
      summary: Create a course and queue content provisioning
      responses: {'200': {description: Course in PROVISIONING state}, '403': {description: ADMIN role required}}
  /api/v1/admin/courses/{courseId}/materials:
    get:
      summary: List course materials through the Authoring facade
      parameters: [{$ref: '#/components/parameters/courseId'}]
      responses: {'200': {description: Materials}, '403': {description: ADMIN role required}}
    post:
      summary: Upload a course material through the Authoring facade
      parameters:
        - {$ref: '#/components/parameters/courseId'}
        - {in: header, name: Idempotency-Key, schema: {type: string}}
      responses: {'200': {description: Material}, '403': {description: ADMIN role required}}
  /api/v1/admin/courses/{courseId}:
    get:
      summary: Read a course for administration
      parameters: [{$ref: '#/components/parameters/courseId'}]
      responses: {'200': {description: Course details}}
    patch:
      summary: Update course publication and metadata
      parameters: [{$ref: '#/components/parameters/courseId'}]
      responses: {'200': {description: Updated course}}
  /api/v1/admin/courses/{courseId}/learning-outcomes:
    get:
      summary: List course learning outcomes
      parameters: [{$ref: '#/components/parameters/courseId'}]
      responses: {'200': {description: Learning outcomes}}
    put:
      summary: Replace course learning outcomes
      parameters: [{$ref: '#/components/parameters/courseId'}]
      responses: {'200': {description: Learning outcomes}}
  /api/v1/admin/courses/{courseId}/mcp-bindings:
    get:
      summary: List read-only MCP bindings for a course
      parameters: [{$ref: '#/components/parameters/courseId'}]
      responses: {'200': {description: MCP bindings}}
    put:
      summary: Replace bindings using discovered read-only tools
      parameters: [{$ref: '#/components/parameters/courseId'}]
      responses: {'200': {description: MCP bindings}}
  /api/v1/admin/courses/{courseId}/materials/{materialId}/retry:
    post:
      summary: Explicitly retry parsing a failed material
      parameters: [{$ref: '#/components/parameters/courseId'}, {$ref: '#/components/parameters/materialId'}]
      responses: {'200': {description: Material}}
  /api/v1/admin/courses/{courseId}/materials/{materialId}/reindex:
    post:
      summary: Explicitly queue material reindexing
      parameters: [{$ref: '#/components/parameters/courseId'}, {$ref: '#/components/parameters/materialId'}]
      responses: {'200': {description: Accepted}}
  /api/v1/admin/courses/{courseId}/materials/{materialId}:
    delete:
      summary: Soft-delete a course material
      parameters: [{$ref: '#/components/parameters/courseId'}, {$ref: '#/components/parameters/materialId'}]
      responses: {'200': {description: Deleted}}
...
  /api/v1/admin/models/{modelId}:
    patch:
      summary: Update an encrypted model profile
      parameters: [{$ref: '#/components/parameters/modelId'}]
      responses: {'200': {description: Model profile with masked credential hint}}
    delete:
      summary: Delete an inactive model profile
      parameters: [{$ref: '#/components/parameters/modelId'}]
      responses: {'200': {description: Deleted}}
  /api/v1/admin/models/{modelId}/test:
    post:
      summary: Test a model profile
      parameters: [{$ref: '#/components/parameters/modelId'}]
      responses: {'200': {description: Test result}}
  /api/v1/admin/models/{modelId}/activate:
    post:
      summary: Activate a model profile for future review snapshots
      parameters: [{$ref: '#/components/parameters/modelId'}]
      responses: {'200': {description: Active model profile}}
...
  /api/v1/admin/mcp/servers/{serverId}:
    put:
      summary: Update an MCP server without exposing secrets
      parameters: [{$ref: '#/components/parameters/serverId'}]
      responses: {'200': {description: MCP server}}
    delete:
      summary: Delete an MCP server
      parameters: [{$ref: '#/components/parameters/serverId'}]
      responses: {'204': {description: Deleted}}
  /api/v1/admin/mcp/servers/{serverId}/test:
    post:
      summary: Validate and discover MCP tools
      parameters: [{$ref: '#/components/parameters/serverId'}]
      responses: {'200': {description: Refreshed MCP server}}
  /api/v1/admin/mcp/servers/{serverId}/refresh:
    post:
      summary: Refresh MCP health and discovered tools
      parameters: [{$ref: '#/components/parameters/serverId'}]
      responses: {'200': {description: Refreshed MCP server}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    courseId: {in: path, name: courseId, required: true, schema: {type: string, format: uuid}}
...
    materialId: {in: path, name: materialId, required: true, schema: {type: string, format: uuid}}
    modelId: {in: path, name: modelId, required: true, schema: {type: string, format: uuid}}
    serverId: {in: path, name: serverId, required: true, schema: {type: string}}
...